+++ src/helix/endpoints/channels/mod.rs
@@ -29,7 +29,9 @@ use std::borrow::Cow;
 
 pub mod add_channel_vip;
 pub mod get_channel_editors;
+pub mod get_channel_followers;
 pub mod get_channel_information;
+pub mod get_followed_channels;
 pub mod get_vips;
 pub mod modify_channel_information;
 pub mod remove_channel_vip;
@@ -40,7 +42,11 @@ pub use add_channel_vip::{AddChannelVipRequest, AddChannelVipResponse};
 #[doc(inline)]
 pub use get_channel_editors::{Editor, GetChannelEditorsRequest};
 #[doc(inline)]
+pub use get_channel_followers::{Follower, GetChannelFollowersRequest};
+#[doc(inline)]
 pub use get_channel_information::{ChannelInformation, GetChannelInformationRequest};
+#[doc(inline)]
+pub use get_followed_channels::{FollowedBroadcaster, GetFollowedChannels};
 #[doc(inline)]
 pub use get_vips::{GetVipsRequest, Vip};
 #[doc(inline)]